Convert class_eval to prepend for models/controllers
This change is done using a convention that allows this class
extensions to be loadable by zeitwerk, included in Rails 6.

This also fixes a load issue since there is a definition of
Spree::Store as a module loaded before this decorator.
