--- app/controllers/spree/api/base_controller_decorator.rb
@@ -1,5 +1,0 @@
-# frozen_string_literal: true
-
-Spree::Api::BaseController.class_eval do
-  include SolidusGlobalize::ControllerGlobalizeHelper
-end
