--- app/models/spree/option_type_decorator.rb
@@ -1,8 +1,0 @@
-# frozen_string_literal: true
-
-module Spree
-  OptionType.class_eval do
-    translates :name, :presentation, fallbacks_for_empty_translations: true
-    include SolidusGlobalize::Translatable
-  end
-end
