--- app/models/spree/promotion_decorator.rb
@@ -1,8 +1,0 @@
-# frozen_string_literal: true
-
-module Spree
-  Promotion.class_eval do
-    translates :name, :description, fallbacks_for_empty_translations: true
-    include SolidusGlobalize::Translatable
-  end
-end
