+++ app/models/spree/store_decorator.rb
@@ -1,9 +1,15 @@
 # frozen_string_literal: true
 
 module Spree
-  Store.class_eval do
-    translates :name, :meta_description, :meta_keywords, :seo_title,
-               fallbacks_for_empty_translations: true
-    include SolidusGlobalize::Translatable
+  module StoreDecorator
+    def self.prepended(base)
+      base.class_eval do
+        translates :name, :meta_description, :meta_keywords, :seo_title,
+                   fallbacks_for_empty_translations: true
+        include SolidusGlobalize::Translatable
+      end
+    end
+
+    ::Spree::Store.prepend self
   end
 end
